+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# [1.72.0](https://github.com/fapdev77/crypto-dashboard/compare/v1.71.0...v1.72.0) (2026-09-20)
+
+
+### Features
+
+* support net balance equity for Bitget and OKX ([25f22bd](https://github.com/fapdev77/crypto-dashboard/commit/25f22bdac5c5e4356124fd422459d07033212884))
+
+# [1.71.0](https://github.com/fapdev77/crypto-dashboard/compare/v1.70.0...v1.71.0) (2026-09-20)
+
+
+### Features
+
+* include account type in filter options ([6e61b27](https://github.com/fapdev77/crypto-dashboard/commit/6e61b270c341976958fe8815f50088682edc8cf0))
+
 # [1.70.0](https://github.com/fapdev77/crypto-dashboard/compare/v1.69.0...v1.70.0) (2026-09-20)